*** empty log message ***

+++ src/engine/rendergl.cpp
@@ -701,7 +701,7 @@ void drawcrosshair(int w, int h)
         if((windowhit ? cursor : crosshair)->bpp==32) glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
         else glBlendFunc(GL_ONE, GL_ONE);
         glColor3f(1, 1, 1);
-        float chsize = (float)(windowhit ? cursorsize : crosshairsize)*w/300;
+        float chsize = (windowhit ? cursorsize : crosshairsize)*w/300.0f;
         float x = w*1.5f - (windowhit ? 0 : chsize/2.0f);
         float y = h*1.5f - (windowhit ? 0 : chsize/2.0f);
         glBindTexture(GL_TEXTURE_2D, (windowhit ? cursor : crosshair)->gl);
